+++ src/demo/components/layout/header/Header.tsx
@@ -1,6 +1,8 @@
 import React from "react";
+import Link from "next/link";
 
-import { Header } from "../../../../index.js";
+import { Header, ICON, StickyHeader } from "../../../../index.js";
+import { HeaderIconAction } from "../../../../components/layout/header/HeaderIconAction";
 
 const HeaderDemo = () => {
     // eslint-disable-next-line @typescript-eslint/no-magic-numbers,react/no-array-index-key
@@ -14,4 +16,96 @@ const HeaderDemo = () => {
     );
 };
 
-export { HeaderDemo };
+const CenteredHeaderDemo = () => {
+    return (
+        <>
+            <Header center={true}>Messages</Header>
+            <div>
+                Header text is centered
+            </div>
+        </>
+    );
+};
+
+const handleClick = () => { alert(1); };
+
+// eslint-disable-next-line max-lines-per-function
+const HeaderWithButtonsDemo = () => {
+    const demo = (
+        <>
+            <HeaderIconAction icon={ICON.back} onClick={handleClick} />
+            <HeaderIconAction icon={ICON.checkmark} onClick={handleClick} />
+        </>
+    );
+
+    return (
+        <>
+            <Header center={true} before={"A text"}>Messages</Header>
+            <div>
+                Header text is centered
+            </div>
+            <Header before={"A text"}>Messages</Header>
+            <div>
+                Header text is start aligned
+            </div>
+            <Header before={demo}>Messages</Header>
+            <div>
+                Some button icons are added on the left to left aligned title
+            </div>
+            <Header before={demo} center={true}>Messages</Header>
+            <div>
+                Some button icons are added on the left to centered aligned title
+            </div>
+            <Header
+                before={<HeaderIconAction icon={ICON.back} to={"/react-miui"} Link={Link} />}
+                center={true}
+            >Messages
+            </Header>
+            <div>
+                Some internal link icons are added on the left to centered aligned title
+            </div>
+            <Header
+                before={<HeaderIconAction icon={ICON.back} href={"https://www.npmjs.com/package/react-miui"} />}
+                center={true}
+            >Messages
+            </Header>
+            <div>
+                Some external link icons are added on the left to centered aligned title
+            </div>
+            <Header
+                before={<HeaderIconAction icon={ICON.back} href={"https://www.npmjs.com/package/react-miui"} />}
+                after={<HeaderIconAction icon={ICON.checkmark} href={"https://www.npmjs.com/package/react-miui"} />}
+                center={true}
+            >Messages
+            </Header>
+            <div>
+                Some external link icons are added on both sides to centered aligned title
+            </div>
+        </>
+    );
+};
+
+const HeaderWithButtonsOnSideDemo = () => {
+    const demo = (
+        <>
+            <HeaderIconAction icon={ICON.back} onClick={handleClick} />
+            <HeaderIconAction icon={ICON.checkmark} onClick={handleClick} />
+        </>
+    );
+
+    return (
+        <StickyHeader position={"left"}>
+            <Header
+                center={true}
+                before={demo}
+                after={demo}
+            >M
+            </Header>
+            <StickyHeader.Content>
+                Header is on the side
+            </StickyHeader.Content>
+        </StickyHeader>
+    );
+};
+
+export { HeaderDemo, CenteredHeaderDemo, HeaderWithButtonsDemo, HeaderWithButtonsOnSideDemo };