+++ src/pages/react-miui.tsx
@@ -1,0 +1,12 @@
+import React from "react";
+
+const ReactMiuiPage = () => {
+    // @TODO nicer UI, add back button etc
+    return (
+        <div>
+            Hi. This is an example page to demonstrate application level routing.
+        </div>
+    );
+};
+
+export default ReactMiuiPage;