+++ src/types.ts
@@ -1,6 +1,6 @@
 import type React from "react";
 
-type AnyComponent = React.ComponentClass | React.FC;
+type AnyComponent = React.ComponentClass | React.FC | React.VFC;
 
 export type {
     AnyComponent,